requirements.txt
requests
python-dotenv
pandas
google-generativeai
phidata
pytest
ratelimit